CHNUKOMBAT/requirements.txt
Flask
Flask-SQLAlchemy
python-telegram-bot